+++ sort.h
@@ -72,5 +72,5 @@ namespace sort {
     }
 
 
-    __device__ void radixSort(unsigned int indices[], unsigned int keys[], unsigned int arrLength, unsigned int keyLength);
+    __device__ void radixSort(unsigned int indices[], uint64_t keys[], unsigned int arrLength, unsigned int keyLength);
 }
